Review fixes, phase 1

+++ segmentation/src/FittedCurve.cpp
@@ -3,7 +3,6 @@
 #include <algorithm>
 #include <cassert>
 #include <cmath>
-#include <iostream>
 
 #include "vc/segmentation/lrps/Derivative.hpp"
 
@@ -83,7 +82,7 @@ auto FittedCurve::sample(const std::size_t numPoints) const
 
 auto FittedCurve::operator()(const int index) const -> Voxel
 {
-    assert(index >= 0 && index < ts_.size() && "out of bounds");
+    assert(index >= 0 && index < static_cast<int>(ts_.size()) && "out of bounds");
     auto p = spline_(ts_[index]);
     return {p(0), p(1), static_cast<double>(zIndex_)};
 }
